requirements.txt
Flask==3.1.1
gspread==6.0.2
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gunicorn==21.2.0
